Move estimateRelayerFee from sdkbase to utils
