+++ packages/utils/src/peripherals/index.ts
@@ -2,3 +2,4 @@ export * from "./chainData";
 export * from "./subgraph";
 export * from "./multipath";
 export * from "./gelato";
+export * from "./relayer";
